\name{AssocAFC}
\alias{AssocAFC}

\docType{package}
\title{
\packageTitle{AssocAFC}
}
\description{
\packageDescription{AssocAFC}
}

\author{
\packageAuthor{AssocAFC}

Maintainer: \packageMaintainer{AssocAFC}
}
\references{
Mohamad Saad, Ellen M. Wijsman; Association score testing for rare variants and binary traits in family data with shared controls, Briefings in Bioinformatics, , bbx107, https://doi.org/10.1093/bib/bbx107
}
\seealso{
\code{\link{Wcorrected}}
\code{\link{Wqls}}
\code{\link{afcSKAT}}
}
\keyword{AssocAFC}
\keyword{package}
\keyword{association}
\keyword{rare}
\keyword{test}
\keyword{family}
\keyword{GWAS}
