\name{cor.afc}
\alias{cor.afc}
\docType{data}
\title{
AFC Correlation Example
}
\description{
An example correlation matrix that corresponds to the the one derived in the unrun example's
first iteration.
}
\usage{data("cor.afc")}
\format{
  The format is:
 num [1:75, 1:75] 1 0.2153 0.14829 -0.00285 0.0536 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:75] "rs3745120_1" "chr19.22817596_1" "chr19.22817734_1" "chr19.22817955_1" ...
  ..$ : chr [1:75] "rs3745120_1" "chr19.22817596_1" "chr19.22817734_1" "chr19.22817955_1" ...
}
\source{
This is simulated data.
}
\examples{
data("cor.afc")
}
\keyword{datasets}
