\name{geno.afc}
\alias{geno.afc}
\docType{data}
\title{
AFC Genotype Example
}
\description{
An example genotype file that corresponds to the the one derived in the unrun example's
first iteration.
}
\usage{data("geno.afc")}
\format{
  The format is:
 int [1:1800, 1:75] 0 0 0 0 0 0 0 0 0 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:75] "rs3745120_1" "chr19.22817596_1" "chr19.22817734_1" "chr19.22817955_1" ...
}
\source{
This is simulated data.
}
\examples{
data("geno.afc")
}
\keyword{datasets}
