\name{kin.afc}
\alias{kin.afc}
\docType{data}
\title{
AFC Kinship Example
}
\description{
An example kinship matrix that corresponds to the the one derived in the unrun example's
first iteration.
}
\usage{data("kin.afc")}
\format{
  The format is:
 num [1:1800, 1:1800] 0.5 0 0.25 0.25 0.25 0.25 0.25 0.25 0.25 0.25 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:1800] "1.1" "1.2" "1.3" "1.4" ...
  ..$ : chr [1:1800] "1.1" "1.2" "1.3" "1.4" ...
}
\source{
This is simulated data.
}
\examples{
data("kin.afc")
}
\keyword{datasets}
