\name{maf.afc}
\alias{maf.afc}
\docType{data}
\title{
AFC MAF Example
}
\description{
An example MAF (Mean Allele Frequency) file that corresponds to the the one derived in the unrun
example's first iteration.
}
\usage{data("maf.afc")}
\format{
  The format is:
 num [1:75, 1:2] 0.00874 0.00541 0.00333 0 0.00291 ...
}
\source{
This is simulated data.
}
\examples{
data("maf.afc")
}
\keyword{datasets}
