\name{phenotype.afc}
\alias{phenotype.afc}
\docType{data}
\title{
AFC Phenotype Example
}
\description{
An example phenotype file that corresponds to the the one derived in the unrun example's
first iteration.
}
\usage{data("phenotype.afc")}
\format{
  The format is:
 int [1:1800, 1] 1 1 1 0 0 1 1 1 1 1 ...
}
\source{
This is simulated data.
}
\examples{
data("phenotype.afc")
}
\keyword{datasets}
