\name{weights.afc}
\alias{weights.afc}
\docType{data}
\title{
AFC Weights Example
}
\description{
An example weights file that corresponds to the the one derived in the unrun example's
first iteration.
}
\usage{data("weights.afc")}
\format{
  The format is:
 num [1:75, 1] 0.996 0.995 0.995 0.999 0.998 ...
}
\source{
This is simulated data.
}
\examples{
data("weights.afc")
}
\keyword{datasets}
